requirements.txt
google-generativeai
python-dotenv
google-cloud-aiplatform
Flask
